feat: migrate to Decimal precision and implement Order Resilience (v0.83.0)
- Migrated all financial calculations to rust_decimal::Decimal

- Implemented OrderMonitor for limit order timeouts

- Integrated Cancel & Replace retry logic in Executor

- Fixed 100+ compilation errors and verified with tests

- Optimized enum memory layout and fixed clippy warnings

+++ src/domain/trading/types.rs
@@ -51,7 +51,7 @@ pub struct Candle {
     pub high: Decimal,
     pub low: Decimal,
     pub close: Decimal,
-    pub volume: f64,
+    pub volume: Decimal,
     pub timestamp: i64,
 }
